+++ CHANGELOG.md
@@ -11,7 +11,7 @@
 - `Repeat@FlowType_K` を追加
 - プロパティメニューに`プロパティ名をコピー`を追加
 - `Motion@FlowType_K` に `Offset` を追加
-- `Kerning@FlowType_K` に `Effect::Parameters` を追加
+- `Kerning@FlowType_K` に `Limit Effects Below` を追加
 - `Blink@FlowType_K` に `Based On` を追加
 - `Island@FlowType_K` で `Time Offset::Order` によって個別オブジェクトのインデックス付与順が変わるように変更
 - キャッシュ処理の修正